Attachment Management (Media Management or file Upload Download) functionality added
Mostly new files. Touch points are in HttpAgent and Service. CommonData and ServiceData are enhanced to carry media information. tested through an example/tutorial page. Only local disk-based storage is available. RDBMS based storage of attachments, as well as interface with corporate central storage systems are to be developed and tested on a need basis