BLOB and CLOB support added
This is primarily accomplished by adding ClobDataType, BlobDataType, ClobValue, BlobVaue classes. Primary design is to use FileManager to create a staging area where the CLOB/BLOB is saved as file during save/retrieve operation. For example if description is a CLOB field in RDBMS, we extract the CLOb field into a temp file, and set the field value (in service context) to this key. This key is sent to client, and the actual file can be streamed to client using Streaming URL.

This opportunity is also used to re-factor few classes around. Utility methods in DbDriver have been made static

This version also has some classes that are getting ready for a revised version of attachment management when the attachments are not stored as CLOB/BLOB. These are not yet wired to the core functionality.