Attachment Management Redesigned and Simplified
fileName and mediaType were managed by us that added to complexity, but unlikely to be used. hence we removed that. Once that burden is removed, saving just the file has simplified the design. Instead of Upload/DownLoad being actions, it is just a declaration of the fields in inputData and OutputData now.